feat(service、dao、util、datasource、orm):新增用户登录接口、新增字符串工具类和集合工具类、修改数据库连接的关闭方法，修改查询多行数据的方法

+++ src/main/java/com/pengwei/www/entity/po/UserRole.java
@@ -21,7 +21,7 @@ public class UserRole {
     /**
      * 角色id
      */
-    private Long roleId;
+    private Integer roleId;
 
     public UserRole() {
     }
@@ -42,11 +42,11 @@ public class UserRole {
         this.userId = userId;
     }
 
-    public Long getRoleId() {
+    public Integer getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Long roleId) {
+    public void setRoleId(Integer roleId) {
         this.roleId = roleId;
     }
 }
